+++ integration-tests/tools/com.liferay.ide.maven.core.tests/src/com/liferay/ide/maven/core/tests/NewModuleMavenTests.java
@@ -17,6 +17,8 @@ package com.liferay.ide.maven.core.tests;
 import com.liferay.ide.maven.core.tests.base.NewModuleMavenBase;
 import com.liferay.ide.project.core.modules.NewLiferayModuleProjectOp;
 
+import org.eclipse.sapphire.Value;
+
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -60,9 +62,13 @@ public class NewModuleMavenTests extends NewModuleMavenBase {
 		op.setProjectProvider(provider());
 		op.setProjectTemplateName("content-targeting-report");
 
-		createOrImportAndBuild(op, project.getName());
+		Value<String> version = op.getLiferayVersion();
 
-		deleteProject(project.getName());
+		if (!"7.2".equals(version.getDefaultContent())) {
+			createOrImportAndBuild(op, project.getName());
+
+			deleteProject(project.getName());
+		}
 	}
 
 	@Test
@@ -73,9 +79,13 @@ public class NewModuleMavenTests extends NewModuleMavenBase {
 		op.setProjectProvider(provider());
 		op.setProjectTemplateName("content-targeting-rule");
 
-		createOrImportAndBuild(op, project.getName());
+		Value<String> version = op.getLiferayVersion();
 
-		deleteProject(project.getName());
+		if (!"7.2".equals(version.getDefaultContent())) {
+			createOrImportAndBuild(op, project.getName());
+
+			deleteProject(project.getName());
+		}
 	}
 
 	@Test
@@ -86,9 +96,13 @@ public class NewModuleMavenTests extends NewModuleMavenBase {
 		op.setProjectProvider(provider());
 		op.setProjectTemplateName("content-targeting-tracking-action");
 
-		createOrImportAndBuild(op, project.getName());
+		Value<String> version = op.getLiferayVersion();
 
-		deleteProject(project.getName());
+		if (!"7.2".equals(version.getDefaultContent())) {
+			createOrImportAndBuild(op, project.getName());
+
+			deleteProject(project.getName());
+		}
 	}
 
 	@Test
@@ -112,9 +126,13 @@ public class NewModuleMavenTests extends NewModuleMavenBase {
 		op.setProjectProvider(provider());
 		op.setProjectTemplateName("form-field");
 
-		createOrImportAndBuild(op, project.getName());
+		Value<String> version = op.getLiferayVersion();
 
-		deleteProject(project.getName());
+		if (!"7.2".equals(version.getDefaultContent())) {
+			createOrImportAndBuild(op, project.getName());
+
+			deleteProject(project.getName());
+		}
 	}
 
 	@Test
